% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_check.R
\name{bivariate_check}
\alias{bivariate_check}
\title{Bivariate Check}
\usage{
bivariate_check(data, alpha)
}
\arguments{
\item{data}{A dataframe}

\item{alpha}{The alpha level you want your ellipse to be}
}
\value{
a list of the proportion of points that lies within the ellipse.
}
\description{
This function takes in multivariate data and an alpha level and makes confidence ellipses depending on the alpha chosen.
Additonally, it returns a list of proportions that lie within the list.
}
\examples{
data = LAGG4793::project_data
bivariate_check(data, 0.5)
}
