% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_plot.R
\name{chisq_plot}
\alias{chisq_plot}
\title{chisq_plot}
\usage{
chisq_plot(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
this function takes in a data frame and make a chi-squared quantile plot. The function also runs a shapiro-wilk test
on the data that has been transformed and determines if the data is normal
}
\examples{
data = LAGG4793::project_data
chisq_plot(data)
}
