% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_chart.R
\name{control_chart}
\alias{control_chart}
\title{control chart}
\usage{
control_chart(data, y_name)
}
\arguments{
\item{data}{a vector of data}

\item{y_name}{the name of the data}
}
\value{
plots an extraordinary event chart
}
\description{
This function takes in data and the name of the data, and it makes an extraordinary event chart
}
\examples{
data = LAGG4793::project_data
data = data$Voltage
control_chart(data, "voltage")
}
