% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonferroni_lambda_ci.R, R/lambda_ci.R
\name{lambda_ci}
\alias{lambda_ci}
\title{lambda_ci}
\usage{
lambda_ci(data, dimensions = 1, alpha = 0.05)

lambda_ci(data, dimensions = 1, alpha = 0.05)
}
\arguments{
\item{data-}{the data}

\item{dimensions-}{the number of variance that are going to be investigated}

\item{alpha-}{the alpha level}
}
\value{
a list containing the two confidence intervals

a list containing the two confidence intervals
}
\description{
This function takes in data, the number of variances interestd in looking at, and an alpha
level to make a (100-alpha)% bonferroni and Anderson and Girshick confidence interval for each variance

This function takes in data, the number of variances interestd in looking at, and an alpha
level to make a (100-alpha)% bonferroni and Anderson and Girshick confidence interval for each variance
}
\examples{
df = LAGG4793::t.84
bonferroni_lambda_ci(df, dimensions = 3, alpha = 0.1)
df = LAGG4793::t.84
bonferroni_lambda_ci(df, dimensions = 3, alpha = 0.1)
}
