% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootfunctions.R
\name{lambda_estimation_cor}
\alias{lambda_estimation_cor}
\title{lambda_estimation_cor}
\usage{
lambda_estimation_cor(data, indices)
}
\arguments{
\item{data-}{the data}

\item{indices-}{indices for bootstrap}
}
\value{
- bootstrap lambda value
}
\description{
lambda_estimation_cor
}
