% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootfunctions.R
\name{lambda_estimation_cov}
\alias{lambda_estimation_cov}
\title{lambda_estimation_cov}
\usage{
lambda_estimation_cov(data, indices)
}
\arguments{
\item{data-}{the data}

\item{indices-}{indices for bootstrap}
}
\value{
- bootstrap lambda value
}
\description{
lambda_estimation_cov
}
