% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_app.R
\name{multivariate_app}
\alias{multivariate_app}
\title{Multivariate Shiny App}
\usage{
multivariate_app()
}
\description{
This is the multivariate shiny app that contains 6 widgets and 3 graphs that show various results for multivariate data.
The full information can be found in the multivariate vignette.
}
\examples{
'multivariate_app()'

}
