% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybootpca.R
\name{mybootpca}
\alias{mybootpca}
\title{mybootpca}
\usage{
mybootpca(data, alpha, iter = 100, cov = TRUE, e_vec = 1)
}
\arguments{
\item{data-}{Data for boot strap}

\item{alpha-}{alpha level}

\item{iter-}{number of iterations}

\item{cov-}{covariance}

\item{p-}{eigen vector}
}
\value{
a list with the confidence intervals for Anderson and Girshick and the bootstrap confidence interval
}
\description{
This function takes in data and makes bootstrap estimates for eigen vectors and values and for Anderson and Girshick confidence intervals
}
\examples{
df = LAGG4793::project_data
df$X = NULL
LAGG4793::mybootpca(df, 0.05)
}
