% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_app.R
\name{normality_app}
\alias{normality_app}
\title{Normality Shiny App}
\usage{
normality_app()
}
\description{
This app is for checking normality for data. There are a total of 8 widgets 3 data frames and 7 plots that run various tests on
the normality of data. Full description can be found in the normality vignette.
}
\examples{
'normality_app()'

}
