% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_decomposition.R
\name{PCA_decomposition}
\alias{PCA_decomposition}
\title{PCA_decomposition}
\usage{
PCA_decomposition(df)
}
\arguments{
\item{df-}{a data frame}
}
\value{
A list containing the covariance matrix, a matrix of the PCA components, the total variance, and a correlation
matrix between each component X of each principle component and each principle component
}
\description{
This function takes in data and uses that data to make PCA components
The components include a covariance matrix of the data, a matrix of each of the makeup of each principle component,
the total sample variance of the PCA decomposition, and a covariance matrix between the variables of
the input data and each PCA component.
}
\examples{
df = LAGG4793::t.84
PCA_decomposition(df)
}
