% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\docType{data}
\name{project_data}
\alias{project_data}
\title{Data of automated welding of automobiles}
\format{
A data set with 40 entries and 5 variables:

\describe{
\item{X}{The observation}
\item{Voltage}{The voltage at which the driveshaft is welded in volts}
\item{Current}{The current at which the driveshaft is welded in amps}
\item{Feedspd}{The feed speed the driveshaft is welded in inches per minute}
\item{Gasf}{The gas flow the driveshaft is welded in cubic feet per minute}
...
}
}
\source{
{Johnson, R. A., & Wichern, D. W. (1992). Applied multivariate statistical analysis. Englewood Cliffs, N.J: Prentice Hall.}
}
\usage{
project_data
}
\description{
The data looks at 40 different data points of how a robot welds a driveshaft to an autombile
}
\keyword{datasets}
