% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_ellipse.R
\name{quality_ellipse}
\alias{quality_ellipse}
\title{quality_ellipse}
\usage{
quality_ellipse(alpha, y, x, y_names, x_names)
}
\arguments{
\item{alpha}{the alpha level of the ellipse}

\item{y}{The vector for the y component of the ellipse}

\item{x}{The vector for the x component of the ellipse}

\item{y_names}{The name of the variable of the y component}

\item{x_names}{The name of the variable of the x component}
}
\value{
a (100-alpha)% ellipse
}
\description{
Creates a quality ellipse based on two variables
}
\examples{
data = LAGG4793::project_data
x = data$Voltage
y = data$Current
quality_ellipse(0.5, y, x, "Current", "Voltage")
}
