% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.84.R
\docType{data}
\name{t.84}
\alias{t.84}
\title{The data looks at weekly rates of returns for 5 stocks}
\format{
A data set with 40 entries and 5 variables:

\describe{
\item{JP Morgan}{The observation}
\item{Citibank}{The voltage at which the driveshaft is welded in volts}
\item{Wells Fargo}{The current at which the driveshaft is welded in amps}
\item{Royal Dutch}{The feed speed the driveshaft is welded in inches per minute}
\item{Shell}{The gas flow the driveshaft is welded in cubic feet per minute}
...
}
}
\source{
{Johnson, R. A., & Wichern, D. W. (1992). Applied multivariate statistical analysis. Englewood Cliffs, N.J: Prentice Hall.}
}
\usage{
t.84
}
\description{
The data looks at weekly rates of returns for 5 stocks
}
\keyword{datasets}
