% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{t-test}
\usage{
t_test(data, means, alpha = 0.05)
}
\arguments{
\item{means}{the means that are being tested}

\item{alpha}{- the alpha level we are testing at}

\item{data-}{a data frame}
}
\value{
The function returns a list of various information about the data and about the results of the test.
}
\description{
t-test
}
\examples{
t_test(data, means, alpha = 0.1)
}
