% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tsq_chart.R
\name{Tsq_chart}
\alias{Tsq_chart}
\title{Tsq_chart}
\usage{
Tsq_chart(data)
}
\arguments{
\item{data}{a dataframe}
}
\description{
This function takes in data and plots a T-squared plot
}
\examples{
data = LAGG4793::project_data
Tsq_chart(data)
}
