% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsq_test.R
\name{tsq_test}
\alias{tsq_test}
\title{tsq_test}
\usage{
tsq_test(data, means, alpha = 0.05)
}
\arguments{
\item{data}{a data frame}

\item{means}{the means that are being tested}

\item{alpha}{the alpha level we are testing at}
}
\value{
The function returns a list of various information about the data and about the results of the test.
}
\description{
tsq_test
}
