% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_check.R
\name{univariate_check}
\alias{univariate_check}
\title{univariate check}
\usage{
univariate_check(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
qq plots of each variable with their shapiro-wilk p-value on it
}
\description{
This funciton uses a shapiro-wilk test to see if each variable in a dataset is normal
}
\examples{
data = data.frame(rnorm(10))
LAGG4793::univariate_check(data)
}
