% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootfunctions.R
\name{vec_estimation_cor}
\alias{vec_estimation_cor}
\title{vec_estimation_cor}
\usage{
vec_estimation_cor(data, indices, p)
}
\arguments{
\item{data-}{the data}

\item{indices-}{indices for bootstrap}
}
\value{
- bootstrap eigen vector value
}
\description{
vec_estimation_cor
}
