% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootfunctions.R
\name{vec_estimation_cov}
\alias{vec_estimation_cov}
\title{vec_estimation_cov}
\usage{
vec_estimation_cov(data, indices)
}
\arguments{
\item{data-}{the data}

\item{indices-}{indices for bootstrap}
}
\value{
- bootstrap eigen vector value
}
\description{
vec_estimation_cov
}
