% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_decomposition.R
\name{PCA_decomposition}
\alias{PCA_decomposition}
\title{PCA_decomposition}
\usage{
PCA_decomposition(df)
}
\arguments{
\item{df-}{the data}
}
\value{
a list that contains the covariance matrix for the input data,
a matrix contianing the components of each principle component, the total sample variance, and the correlation matrix
between the x components and the principle components
}
\description{
This function takes in data and returns a list that contains the covariance matrix for the input data,
a matrix contianing the components of each principle component, the total sample variance, and the correlation matrix
between the x components and the principle components
}
\examples{
df = LAGG4793::t.84
PCA_decomposition(df)
}
